requirements.txt
streamlit
pandas
pandas_datareader
datetime
plotly.express